requirements.txt
absl-py==2.1.0
asttokens==3.0.0
attrs==25.3.0
autocommand==2.2.2
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy==1.3.0
cycler==0.12.1
-e git+https://github.com/KaiyangZhou/Dassl.pytorch.git@c61a1b570ac6333bd50fb5ae06aea59002fb20bb#egg=dassl
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
executing==2.2.0
fastjsonschema==2.21.1
filelock @ file:///croot/filelock_1700591183607/work
flake8==3.7.9
fonttools==4.55.4
fsspec==2024.10.0
ftfy==6.3.0
future==1.0.0
gdown==5.2.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
grad-cam==1.5.5
grpcio==1.67.0
huggingface-hub==0.27.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.4.0
inflect==7.3.1
ipython==8.12.3
isort==4.3.21
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
littleutils==0.2.4
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.6.1
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
more-itertools==10.3.0
mpmath @ file:///croot/mpmath_1690848262763/work
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.23.5
ogb==1.3.6
opencv-python==4.11.0.86
outdated==0.2.2
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.51
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.4.0
referencing==0.36.2
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
rpds-py==0.25.1
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tb-nightly==2.19.0a20241022
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.14
tinycss2==1.4.0
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
triton==2.2.0
ttach==0.0.3
typeguard==4.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
yarg==0.1.9
zipp==3.20.2
